refactor(views): migrate queue to per-folder layout (commit 7/7)
Split src/views/queue.rs (1802 lines) into src/views/queue/{mod,update,view}.rs.
Queue is the outlier of the family — uses QueueSortMode (not SortMode),
does not implement HasCommonAction, and carries queue-specific behaviors
(drag-and-drop reorder, playlist edit-mode bar, quick-save context bar)
not shared with the library views.

- mod.rs: types (Page, Message, Action, ViewData, QueueContextEntry,
  re-export of QueueSortMode), Default + new(), explicit ViewPage
  impl (no sort_mode_selected_message override since queue's sort uses
  QueueSortMode), tests for column visibility default + get/set round-trip
- update.rs: impl QueuePage { fn update } + the toggle-column test
- view.rs: impl QueuePage { fn view } + the BREAKPOINT_HIDE_QUEUE_STARS
  constant + six column-visibility helpers (rating with width-responsive,
  album / duration / love / plays / genre) + 10 helper tests covering
  the responsive width gate, sort-mode-driven auto-show, and user-toggle
  precedence

Macro consumers:
- define_view_columns! for QueueColumn / Visibility (9 columns).
- impl_has_common_action!: NOT used — queue does not implement
  HasCommonAction. The macro's no_center variant exists for the
  Playlists case; queue's lack of any common-action impl is its own
  thing entirely.

Final commit of the views-unification refactor. All six monolithic view
files are now per-folder modules following the proven views/settings/
template.

cargo build / clippy --all-targets -D warnings / nightly fmt --check /
test --workspace all clean (662 + 368 + 1 doctest).
