+++ src/views/queue/view.rs
@@ -1,115 +1,19 @@
-//! Queue Page Component
+//! Queue view — `impl QueuePage { fn view }`.
 //!
-//! Self-contained queue view with slot list navigation showing currently playing queue.
-//! Uses message bubbling pattern to communicate global actions to root.
+//! Rendering for the queue page, plus the per-mode column-visibility helpers
+//! and the `BREAKPOINT_HIDE_QUEUE_STARS` constant. Update/state logic lives
+//! in `update.rs`; types live in `mod.rs`.
 
 use iced::{
-    Alignment, Element, Length, Task,
+    Alignment, Element, Length,
     widget::{Row, button, column, container, mouse_area, row},
 };
-// Re-export QueueSortMode from data crate (canonical definition)
 use nokkvi_data::backend::queue::QueueSongUIViewData;
-pub(crate) use nokkvi_data::types::queue_sort_mode::QueueSortMode;
-use tracing::{debug, trace};
 
-use crate::{
-    app_message::Message,
-    widgets::{self, SlotListPageState, drag_column::DragEvent, hover_overlay::HoverOverlay},
+use super::{
+    QueueColumn, QueueContextEntry, QueueMessage, QueuePage, QueueSortMode, QueueViewData,
 };
-
-/// Queue page local state
-#[derive(Debug)]
-pub struct QueuePage {
-    pub common: SlotListPageState,
-    /// Queue uses its own sort mode enum (QueueSortMode), separate from
-    /// the library views' SortMode.
-    pub queue_sort_mode: QueueSortMode,
-    /// Per-column visibility toggles surfaced via the columns-3-cog dropdown
-    /// in the view header. Persisted to config.toml.
-    pub column_visibility: QueueColumnVisibility,
-    /// Cache of the last `(mode, ascending, queue_len)` that was applied. The
-    /// queue sort short-circuits when this matches — covers the common
-    /// "user toggles same sort mode again" case and most "queue length
-    /// unchanged since last sort" cases. Same-length-different-content
-    /// requires the caller to manually re-trigger or invalidate.
-    pub last_sort_signature: Option<(QueueSortMode, bool, usize)>,
-}
-
-/// Toggleable queue columns. `Stars`, `Album`, `Duration`, `Love`, and `Plays`
-/// are user-toggleable from the columns dropdown; the index/title/artwork
-/// columns stay always-on.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum QueueColumn {
-    Select,
-    Index,
-    Thumbnail,
-    Stars,
-    Album,
-    Duration,
-    Love,
-    Plays,
-    Genre,
-}
-
-/// User-toggle state for each toggleable queue column.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct QueueColumnVisibility {
-    pub select: bool,
-    pub index: bool,
-    pub thumbnail: bool,
-    pub stars: bool,
-    pub album: bool,
-    pub duration: bool,
-    pub love: bool,
-    pub plays: bool,
-    pub genre: bool,
-}
-
-impl Default for QueueColumnVisibility {
-    fn default() -> Self {
-        Self {
-            select: false,
-            index: true,
-            thumbnail: true,
-            stars: true,
-            album: true,
-            duration: true,
-            love: true,
-            plays: false,
-            genre: false,
-        }
-    }
-}
-
-impl QueueColumnVisibility {
-    pub fn get(&self, col: QueueColumn) -> bool {
-        match col {
-            QueueColumn::Select => self.select,
-            QueueColumn::Index => self.index,
-            QueueColumn::Thumbnail => self.thumbnail,
-            QueueColumn::Stars => self.stars,
-            QueueColumn::Album => self.album,
-            QueueColumn::Duration => self.duration,
-            QueueColumn::Love => self.love,
-            QueueColumn::Plays => self.plays,
-            QueueColumn::Genre => self.genre,
-        }
-    }
-
-    pub fn set(&mut self, col: QueueColumn, value: bool) {
-        match col {
-            QueueColumn::Select => self.select = value,
-            QueueColumn::Index => self.index = value,
-            QueueColumn::Thumbnail => self.thumbnail = value,
-            QueueColumn::Stars => self.stars = value,
-            QueueColumn::Album => self.album = value,
-            QueueColumn::Duration => self.duration = value,
-            QueueColumn::Love => self.love = value,
-            QueueColumn::Plays => self.plays = value,
-            QueueColumn::Genre => self.genre = value,
-        }
-    }
-}
+use crate::widgets::{self, hover_overlay::HoverOverlay};
 
 /// Hide the queue stars column when the queue panel is narrower than this.
 /// Queue panel is measured (via `iced::widget::responsive`), so this fires
@@ -159,472 +63,7 @@ pub(crate) fn genre_column_visible(sort: QueueSortMode, user_visible: bool) -> b
     user_visible || matches!(sort, QueueSortMode::Genre)
 }
 
-/// View data passed from root (read-only)
-pub struct QueueViewData<'a> {
-    pub queue_songs: std::borrow::Cow<'a, [QueueSongUIViewData]>,
-    pub album_art: &'a std::collections::HashMap<String, iced::widget::image::Handle>,
-    pub large_artwork: &'a std::collections::HashMap<String, iced::widget::image::Handle>,
-    pub window_width: f32,
-    pub window_height: f32,
-    pub scale_factor: f32,
-    pub modifiers: iced::keyboard::Modifiers,
-    pub current_playing_song_id: Option<String>,
-    pub current_playing_queue_index: Option<usize>,
-    pub is_playing: bool, // True if playback is active (not stopped/paused)
-    pub total_queue_count: usize, // Total count before filtering (for empty state detection)
-    pub stable_viewport: bool,
-    /// When in edit mode: (playlist_name, is_dirty)
-    pub edit_mode_info: Option<(String, bool)>,
-    /// Playlist comment when in edit mode
-    pub edit_mode_comment: Option<String>,
-    /// Playlist public flag when in edit mode (drives the lock toggle button)
-    pub edit_mode_public: Option<bool>,
-    /// When a playlist is loaded for playback (not editing)
-    pub playlist_context_info: Option<crate::state::ActivePlaylistContext>,
-    /// Whether the column-visibility checkbox dropdown is open (controlled
-    /// by `Nokkvi.open_menu`).
-    pub column_dropdown_open: bool,
-    /// Trigger bounds captured when the dropdown was opened.
-    pub column_dropdown_trigger_bounds: Option<iced::Rectangle>,
-    /// Borrowed reference to the root open-menu state, so per-row context
-    /// menus can resolve their own open/closed status.
-    pub open_menu: Option<&'a crate::app_message::OpenMenu>,
-    /// Whether the queue's view-header chip should render. Gated by the
-    /// `queue_show_default_playlist` user setting.
-    pub show_default_playlist_chip: bool,
-    /// Current default-playlist display name (empty when no default set).
-    pub default_playlist_name: &'a str,
-}
-
-/// Context menu entries for queue items
-#[derive(Debug, Clone, Copy)]
-pub enum QueueContextEntry {
-    Play,
-    PlayNext,
-    Separator,
-    RemoveFromQueue,
-    AddToPlaylist,
-    SaveAsPlaylist,
-    OpenBrowsingPanel,
-    GetInfo,
-    ShowInFolder,
-    FindSimilar,
-    TopSongs,
-}
-
-/// Messages for local queue page interactions
-#[derive(Debug, Clone)]
-pub enum QueueMessage {
-    // Slot list navigation
-    SlotListNavigateUp,
-    SlotListNavigateDown,
-    SlotListSetOffset(usize, iced::keyboard::Modifiers),
-    SlotListScrollSeek(usize),
-    SlotListActivateCenter,
-    SlotListClickPlay(usize), // Click non-center to play directly (skip focus)
-    /// Click on a row's leading select checkbox — toggles `item_index` in the
-    /// page's `selected_indices` set. No play/highlight side effects.
-    SlotListSelectionToggle(usize),
-    /// Click on the tri-state "select all" header checkbox — fills the
-    /// selection set with every visible (filtered) row, or clears it when
-    /// every visible row is already selected.
-    SlotListSelectAllToggle,
-    FocusCurrentPlaying(usize, bool), // Auto-scroll slot list to center currently playing track (by queue index, flash)
-
-    // Mouse click on star/heart (item_index, value)
-    ClickSetRating(usize, usize), // (item_index, rating 1-5)
-    ClickToggleStar(usize),       // item_index
-
-    // Context menu
-    ContextMenuAction(usize, QueueContextEntry), // (item_index, entry)
-
-    // Drag-and-drop reordering
-    DragReorder(DragEvent),
-
-    // View header interactions
-    SortModeSelected(QueueSortMode),
-    ToggleSortOrder,
-    SearchQueryChanged(String),
-    ToggleColumnVisible(QueueColumn),
-    /// Sort dropdown's "Roulette" entry was selected — intercepted at the
-    /// root handler before the page's `update` runs.
-    Roulette,
-
-    // Playlist edit mode
-    SavePlaylist,
-    DiscardEdits,
-    PlaylistNameChanged(String),
-    PlaylistCommentChanged(String),
-    PlaylistEditPublicToggled(bool),
-    EditPlaylist,      // Enter edit mode for the currently-playing playlist
-    QuickSavePlaylist, // Save current queue back to the active playlist without entering edit mode
-
-    // Data loading (moved from root Message enum)
-    QueueLoaded(Result<Vec<QueueSongUIViewData>, String>), // queue_songs
-    /// Refresh artwork for a specific album (album_id)
-    RefreshArtwork(String),
-    /// Navigate to a view and apply an ID filter
-    NavigateAndFilter(crate::View, nokkvi_data::types::filter::LibraryFilter),
-    /// Navigate to Albums and auto-expand the album with this id (no filter set).
-    NavigateAndExpandAlbum(String),
-    /// Navigate to Artists and auto-expand the artist with this id (no filter set).
-    NavigateAndExpandArtist(String),
-    /// Navigate to Genres and auto-expand the genre with this id (no filter set).
-    NavigateAndExpandGenre(String),
-    /// Column-dropdown open/close request — bubbled to root
-    /// `Message::SetOpenMenu`. Intercepted in `handle_queue` before the
-    /// page's `update` runs.
-    SetOpenMenu(Option<crate::app_message::OpenMenu>),
-    /// Artwork column drag handle event — intercepted at root, page never sees it.
-    ArtworkColumnDrag(crate::widgets::artwork_split_handle::DragEvent),
-    /// Header chip clicked — bubble to root, opens the default-playlist picker.
-    OpenDefaultPlaylistPicker,
-}
-
-/// Actions that bubble up to root for global state mutation
-#[derive(Debug, Clone)]
-pub enum QueueAction {
-    PlaySong(usize),                // song index in queue
-    FocusOnSong(usize, bool),       // queue index to scroll to (bubbles up to handler), flash
-    SortModeChanged(QueueSortMode), // trigger reload/resort
-    SortOrderChanged(bool),         // trigger resort
-    SearchChanged(String),          // trigger filter
-    SetRating(String, usize),       // (song_id, rating) - set absolute rating
-    ToggleStar(String, bool),       // (song_id, new_starred) - toggle starred state
-    MoveItem {
-        from: usize,
-        to: usize,
-    }, // drag-and-drop reorder (absolute item indices)
-    MoveBatch {
-        indices: Vec<usize>,
-        target: usize,
-    }, // multi-selection drag reorder
-    RemoveFromQueue(Vec<String>),   // remove songs by ID (immune to index drift)
-    PlayNext(Vec<String>),          // insert songs after currently playing (by ID)
-    ShowToast(String),              // informational toast (e.g. drag disabled reason)
-    SaveAsPlaylist,                 // open dialog to save queue as new playlist
-    OpenBrowsingPanel,              // toggle the library browser panel
-    AddToPlaylist(Vec<String>),     // song_ids - add to playlist dialog
-    SavePlaylist,                   // save playlist edits (edit mode)
-    DiscardEdits,                   // discard edits and exit edit mode
-    PlaylistNameChanged(String),    // playlist name edited inline
-    PlaylistCommentChanged(String), // playlist comment edited inline
-    PlaylistEditPublicToggled(bool), // public/private toggled in the edit bar
-    EditPlaylist,                   // enter edit mode from playlist context bar
-    ShowInfo(usize),                // Open info modal (queue index for full Song lookup)
-    ShowInFolder(usize),            // Open containing folder (queue index, path fetched via API)
-    RefreshArtwork(String),         // album_id - refresh artwork from server
-    FindSimilar(usize),             // Open Find Similar panel for queue index
-    TopSongs(usize),                // Open Top Songs panel for queue index
-    NavigateAndFilter(crate::View, nokkvi_data::types::filter::LibraryFilter), // Navigate to target view and filter
-    NavigateAndExpandAlbum(String), // album_id - navigate to Albums and auto-expand
-    NavigateAndExpandArtist(String), // artist_id - navigate to Artists and auto-expand
-    NavigateAndExpandGenre(String), // genre_id - navigate to Genres and auto-expand
-    /// User toggled a queue column's visibility — persist to config.toml.
-    ColumnVisibilityChanged(QueueColumn, bool),
-    /// Bubble to root: open the default-playlist picker overlay.
-    OpenDefaultPlaylistPicker,
-    None,
-}
-
-impl Default for QueuePage {
-    fn default() -> Self {
-        Self {
-            common: SlotListPageState::new_without_sort_mode(),
-            queue_sort_mode: QueueSortMode::Album,
-            column_visibility: QueueColumnVisibility::default(),
-            last_sort_signature: None,
-        }
-    }
-}
-
 impl QueuePage {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Update internal state and return actions for root
-    pub fn update(
-        &mut self,
-        message: QueueMessage,
-        queue_songs: &[QueueSongUIViewData],
-    ) -> (Task<QueueMessage>, QueueAction) {
-        let total_items = queue_songs.len();
-        match message {
-            QueueMessage::SlotListNavigateUp => {
-                self.common.handle_navigate_up(total_items);
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::SlotListNavigateDown => {
-                self.common.handle_navigate_down(total_items);
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::SlotListSetOffset(offset, modifiers) => {
-                self.common
-                    .handle_slot_click(offset, total_items, modifiers);
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::SlotListScrollSeek(offset) => {
-                self.common.handle_set_offset(offset, total_items);
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::SlotListActivateCenter => {
-                // Play the centered song
-                if let Some(center_idx) = self.common.get_center_item_index(total_items) {
-                    self.common.slot_list.flash_center();
-                    (Task::none(), QueueAction::PlaySong(center_idx))
-                } else {
-                    (Task::none(), QueueAction::None)
-                }
-            }
-            QueueMessage::SlotListClickPlay(offset) => {
-                self.common.handle_set_offset(offset, total_items);
-                self.update(QueueMessage::SlotListActivateCenter, queue_songs)
-            }
-            QueueMessage::SlotListSelectionToggle(offset) => {
-                self.common.handle_selection_toggle(offset, total_items);
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::SlotListSelectAllToggle => {
-                self.common.handle_select_all_toggle(total_items);
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::FocusCurrentPlaying(queue_index, flash) => {
-                // Auto-scroll slot list to center the currently playing track by queue index
-                // Bubble up to handler which has access to queue_songs to find the slot
-                trace!(
-                    " [QUEUE PAGE] FocusCurrentPlaying({}) called, current_offset={}",
-                    queue_index, self.common.slot_list.viewport_offset
-                );
-                (Task::none(), QueueAction::FocusOnSong(queue_index, flash))
-            }
-            QueueMessage::NavigateAndFilter(view, filter) => {
-                (Task::none(), QueueAction::NavigateAndFilter(view, filter))
-            }
-            QueueMessage::NavigateAndExpandAlbum(album_id) => {
-                (Task::none(), QueueAction::NavigateAndExpandAlbum(album_id))
-            }
-            QueueMessage::NavigateAndExpandArtist(artist_id) => (
-                Task::none(),
-                QueueAction::NavigateAndExpandArtist(artist_id),
-            ),
-            QueueMessage::NavigateAndExpandGenre(genre_id) => {
-                (Task::none(), QueueAction::NavigateAndExpandGenre(genre_id))
-            }
-            QueueMessage::SortModeSelected(sort_mode) => {
-                self.queue_sort_mode = sort_mode;
-                (Task::none(), QueueAction::SortModeChanged(sort_mode))
-            }
-            QueueMessage::ToggleSortOrder => {
-                self.common.sort_ascending = !self.common.sort_ascending;
-                (
-                    Task::none(),
-                    QueueAction::SortOrderChanged(self.common.sort_ascending),
-                )
-            }
-            QueueMessage::SearchQueryChanged(query) => {
-                self.common.search_query = query.clone();
-                self.common.slot_list.set_offset(0, total_items); // Reset to top on search
-                (Task::none(), QueueAction::SearchChanged(query))
-            }
-            QueueMessage::ToggleColumnVisible(col) => {
-                let new_value = !self.column_visibility.get(col);
-                self.column_visibility.set(col, new_value);
-                (
-                    Task::none(),
-                    QueueAction::ColumnVisibilityChanged(col, new_value),
-                )
-            }
-
-            // Data loading messages (handled at root level, no action needed here)
-            QueueMessage::QueueLoaded(_) => (Task::none(), QueueAction::None),
-            // Routed up to root in `handle_queue` before this match runs;
-            // arm exists only for exhaustiveness.
-            QueueMessage::SetOpenMenu(_) => (Task::none(), QueueAction::None),
-            QueueMessage::Roulette => (Task::none(), QueueAction::None),
-            QueueMessage::ArtworkColumnDrag(_) => {
-                // Intercepted at root before reaching this update; never reached.
-                (Task::none(), QueueAction::None)
-            }
-            QueueMessage::OpenDefaultPlaylistPicker => {
-                (Task::none(), QueueAction::OpenDefaultPlaylistPicker)
-            }
-            QueueMessage::ClickSetRating(item_index, rating) => {
-                if let Some(song) = queue_songs.get(item_index) {
-                    let current = song.rating.unwrap_or(0) as usize;
-                    let new_rating = if rating == current {
-                        rating.saturating_sub(1)
-                    } else {
-                        rating
-                    };
-                    (
-                        Task::none(),
-                        QueueAction::SetRating(song.id.clone(), new_rating),
-                    )
-                } else {
-                    (Task::none(), QueueAction::None)
-                }
-            }
-            QueueMessage::ClickToggleStar(item_index) => {
-                if let Some(song) = queue_songs.get(item_index) {
-                    (
-                        Task::none(),
-                        QueueAction::ToggleStar(song.id.clone(), !song.starred),
-                    )
-                } else {
-                    (Task::none(), QueueAction::None)
-                }
-            }
-            QueueMessage::DragReorder(drag_event) => {
-                // Drag is allowed in any sort mode, but blocked during active search
-                let drag_allowed = self.common.search_query.is_empty();
-
-                match drag_event {
-                    DragEvent::Picked { .. } if !drag_allowed => (
-                        Task::none(),
-                        QueueAction::ShowToast("Clear search to reorder queue".to_string()),
-                    ),
-                    DragEvent::Dropped {
-                        index,
-                        target_index,
-                    } if drag_allowed => {
-                        // Translate slot indices to absolute item indices using the
-                        // same effective_center logic that build_slot_list_slots uses for
-                        // rendering. Simple `viewport_offset + slot` is wrong because
-                        // it doesn't account for the center_slot offset.
-                        let from = self.common.slot_list.slot_to_item_index(index, total_items);
-                        let to = self
-                            .common
-                            .slot_list
-                            .slot_to_item_index_for_drop(target_index, total_items);
-                        debug!(
-                            "\u{1f4e6} [QUEUE] Drag reorder: slot {}\u{2192}{} \u{2192} item {:?}\u{2192}{:?} \\
-                             (viewport_offset={}, slot_count={}, total={})",
-                            index,
-                            target_index,
-                            from,
-                            to,
-                            self.common.slot_list.viewport_offset,
-                            self.common.slot_list.slot_count,
-                            total_items,
-                        );
-
-                        // Multi-selection batch drag: if selected_indices has multiple
-                        // items and the dragged item is one of them, move the whole batch.
-                        let selected = &self.common.slot_list.selected_indices;
-                        if selected.len() > 1
-                            && from.is_some_and(|f| selected.contains(&f))
-                            && let Some(t) = to
-                        {
-                            let indices: Vec<usize> = selected.iter().copied().collect();
-                            self.common.clear_multi_selection();
-                            (Task::none(), QueueAction::MoveBatch { indices, target: t })
-                        } else {
-                            match (from, to) {
-                                (Some(f), Some(t)) => {
-                                    // Keep highlight on the moved item at its new position
-                                    let insert_at = if f < t { t - 1 } else { t };
-                                    self.common.slot_list.set_selected(insert_at, total_items);
-                                    (Task::none(), QueueAction::MoveItem { from: f, to: t })
-                                }
-                                _ => {
-                                    debug!(
-                                        "\u{1f4e6} [QUEUE] Drag dropped on empty slot, ignoring"
-                                    );
-                                    (Task::none(), QueueAction::None)
-                                }
-                            }
-                        }
-                    }
-                    DragEvent::Picked { index } if drag_allowed => {
-                        // Check if the picked item is part of an active multi-selection.
-                        // If yes, preserve the selection (batch drag). If not, highlight
-                        // only the dragged item (single drag).
-                        if let Some(item_index) =
-                            self.common.slot_list.slot_to_item_index(index, total_items)
-                            && !self.common.slot_list.selected_indices.contains(&item_index)
-                        {
-                            self.common.slot_list.set_selected(item_index, total_items);
-                        }
-                        (Task::none(), QueueAction::None)
-                    }
-                    _ => (Task::none(), QueueAction::None),
-                }
-            }
-            QueueMessage::ContextMenuAction(clicked_idx, entry) => match entry {
-                QueueContextEntry::Play => {
-                    self.common.handle_set_offset(clicked_idx, total_items);
-                    (Task::none(), QueueAction::PlaySong(clicked_idx))
-                }
-                QueueContextEntry::RemoveFromQueue | QueueContextEntry::PlayNext => {
-                    let target_indices = self.common.evaluate_context_menu(clicked_idx);
-                    self.common.clear_multi_selection();
-
-                    // Resolve filtered indices → song IDs at the boundary so
-                    // downstream code is index-free. Stale `track_number`,
-                    // client-side sorts, and optimistic mutations can't desync
-                    // an ID-based target.
-                    let target_ids: Vec<String> = target_indices
-                        .iter()
-                        .filter_map(|&idx| queue_songs.get(idx).map(|s| s.id.clone()))
-                        .collect();
-
-                    match entry {
-                        QueueContextEntry::RemoveFromQueue => {
-                            (Task::none(), QueueAction::RemoveFromQueue(target_ids))
-                        }
-                        QueueContextEntry::PlayNext => {
-                            (Task::none(), QueueAction::PlayNext(target_ids))
-                        }
-                        _ => unreachable!(),
-                    }
-                }
-                QueueContextEntry::AddToPlaylist => {
-                    let target_indices = self.common.evaluate_context_menu(clicked_idx);
-                    self.common.clear_multi_selection();
-                    let target_songs: Vec<String> = target_indices
-                        .iter()
-                        .filter_map(|&idx| queue_songs.get(idx).map(|s| s.id.clone()))
-                        .collect();
-                    if target_songs.is_empty() {
-                        (Task::none(), QueueAction::None)
-                    } else {
-                        (Task::none(), QueueAction::AddToPlaylist(target_songs))
-                    }
-                }
-                QueueContextEntry::Separator => (Task::none(), QueueAction::None),
-                QueueContextEntry::SaveAsPlaylist => (Task::none(), QueueAction::SaveAsPlaylist),
-                QueueContextEntry::OpenBrowsingPanel => {
-                    (Task::none(), QueueAction::OpenBrowsingPanel)
-                }
-                QueueContextEntry::GetInfo => (Task::none(), QueueAction::ShowInfo(clicked_idx)),
-                QueueContextEntry::ShowInFolder => {
-                    (Task::none(), QueueAction::ShowInFolder(clicked_idx))
-                }
-                QueueContextEntry::FindSimilar => {
-                    (Task::none(), QueueAction::FindSimilar(clicked_idx))
-                }
-                QueueContextEntry::TopSongs => (Task::none(), QueueAction::TopSongs(clicked_idx)),
-            },
-            QueueMessage::SavePlaylist => (Task::none(), QueueAction::SavePlaylist),
-            QueueMessage::DiscardEdits => (Task::none(), QueueAction::DiscardEdits),
-            QueueMessage::PlaylistNameChanged(name) => {
-                (Task::none(), QueueAction::PlaylistNameChanged(name))
-            }
-            QueueMessage::PlaylistCommentChanged(comment) => {
-                (Task::none(), QueueAction::PlaylistCommentChanged(comment))
-            }
-            QueueMessage::PlaylistEditPublicToggled(value) => {
-                (Task::none(), QueueAction::PlaylistEditPublicToggled(value))
-            }
-            QueueMessage::EditPlaylist => (Task::none(), QueueAction::EditPlaylist),
-            QueueMessage::QuickSavePlaylist => (Task::none(), QueueAction::SaveAsPlaylist),
-            QueueMessage::RefreshArtwork(album_id) => {
-                (Task::none(), QueueAction::RefreshArtwork(album_id))
-            }
-        }
-    }
-
     /// Build the view
     pub fn view<'a>(&'a self, data: QueueViewData<'a>) -> Element<'a, QueueMessage> {
         use crate::widgets::slot_list::{SlotListConfig, SlotListRowContext};
@@ -1567,31 +1006,6 @@ impl QueuePage {
     }
 }
 
-// ============================================================================
-// ViewPage trait implementation
-// ============================================================================
-
-impl super::ViewPage for QueuePage {
-    fn common(&self) -> &SlotListPageState {
-        &self.common
-    }
-    fn common_mut(&mut self) -> &mut SlotListPageState {
-        &mut self.common
-    }
-
-    fn search_input_id(&self) -> &'static str {
-        super::QUEUE_SEARCH_ID
-    }
-
-    // Queue uses QueueSortMode, not SortMode — sort_mode_selected_message returns None (default).
-    fn toggle_sort_order_message(&self) -> Message {
-        Message::Queue(QueueMessage::ToggleSortOrder)
-    }
-
-    // Queue items are already in the queue, so add_to_queue_message returns None (default).
-    // Queue has no reload_message (client-side filtering, no server fetch needed on Escape).
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1675,39 +1089,6 @@ mod tests {
     }
 
     #[test]
-    fn queue_column_visibility_default_shows_legacy_columns_only() {
-        let v = QueueColumnVisibility::default();
-        assert!(v.stars);
-        assert!(v.album);
-        assert!(v.duration);
-        assert!(v.love);
-        // Plays is opt-in (auto-shown only when sort = MostPlayed).
-        assert!(!v.plays);
-    }
-
-    #[test]
-    fn queue_column_visibility_get_set_round_trip() {
-        let mut v = QueueColumnVisibility::default();
-        assert!(v.get(QueueColumn::Stars));
-        v.set(QueueColumn::Stars, false);
-        assert!(!v.get(QueueColumn::Stars));
-        // Other columns unchanged.
-        assert!(v.get(QueueColumn::Album));
-        assert!(v.get(QueueColumn::Duration));
-        assert!(v.get(QueueColumn::Love));
-        assert!(!v.get(QueueColumn::Plays));
-
-        v.set(QueueColumn::Album, false);
-        v.set(QueueColumn::Duration, false);
-        v.set(QueueColumn::Love, false);
-        v.set(QueueColumn::Plays, true);
-        assert!(!v.get(QueueColumn::Album));
-        assert!(!v.get(QueueColumn::Duration));
-        assert!(!v.get(QueueColumn::Love));
-        assert!(v.get(QueueColumn::Plays));
-    }
-
-    #[test]
     fn plays_column_visible_auto_shows_on_most_played() {
         // Sort overrides the user toggle: MostPlayed always shows, regardless of toggle.
         assert!(plays_column_visible(QueueSortMode::MostPlayed, false));
@@ -1734,69 +1115,5 @@ mod tests {
         assert!(genre_column_visible(QueueSortMode::Title, true));
         assert!(!genre_column_visible(QueueSortMode::MostPlayed, false));
         assert!(genre_column_visible(QueueSortMode::MostPlayed, true));
-    }
-
-    #[test]
-    fn queue_column_visibility_default_keeps_genre_off() {
-        let v = QueueColumnVisibility::default();
-        assert!(!v.genre);
-    }
-
-    #[test]
-    fn toggle_column_visible_flips_state() {
-        let mut page = QueuePage::default();
-        let songs: Vec<QueueSongUIViewData> = Vec::new();
-
-        // Stars: true → false → true.
-        let (_t, action) = page.update(
-            QueueMessage::ToggleColumnVisible(QueueColumn::Stars),
-            &songs,
-        );
-        assert!(!page.column_visibility.stars);
-        assert!(matches!(
-            action,
-            QueueAction::ColumnVisibilityChanged(QueueColumn::Stars, false)
-        ));
-
-        let (_t, action) = page.update(
-            QueueMessage::ToggleColumnVisible(QueueColumn::Stars),
-            &songs,
-        );
-        assert!(page.column_visibility.stars);
-        assert!(matches!(
-            action,
-            QueueAction::ColumnVisibilityChanged(QueueColumn::Stars, true)
-        ));
-
-        // Album and Duration use the same path, just spot-check Album.
-        let (_t, _action) = page.update(
-            QueueMessage::ToggleColumnVisible(QueueColumn::Album),
-            &songs,
-        );
-        assert!(!page.column_visibility.album);
-
-        // Genre default is off → toggle ON, message carries Genre+true.
-        let (_t, action) = page.update(
-            QueueMessage::ToggleColumnVisible(QueueColumn::Genre),
-            &songs,
-        );
-        assert!(page.column_visibility.genre);
-        assert!(matches!(
-            action,
-            QueueAction::ColumnVisibilityChanged(QueueColumn::Genre, true)
-        ));
-        // Other columns unaffected.
-        assert!(page.column_visibility.stars);
-        assert!(page.column_visibility.duration);
-        assert!(page.column_visibility.love);
-
-        // Love toggles independently and emits its own action.
-        let (_t, action) =
-            page.update(QueueMessage::ToggleColumnVisible(QueueColumn::Love), &songs);
-        assert!(!page.column_visibility.love);
-        assert!(matches!(
-            action,
-            QueueAction::ColumnVisibilityChanged(QueueColumn::Love, false)
-        ));
     }
 }
